+++ tests/Feature/NotificationAndDashboardTest.php
@@ -20,7 +20,6 @@ class NotificationAndDashboardTest extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
-        Setting::create(['setting_key' => 'work_hours_enabled', 'value' => '0']);
         Setting::create(['setting_key' => 'site_name', 'value' => '测试工单']);
     }